Fix TypeError: cast correctness to float before subtracting in interaction_ci
results.csv's correctness column comes back as bool; pandas/numpy no longer
allow subtracting two bool Series directly. score_against_run already casts
via .to_numpy(float) elsewhere -- interaction_ci's direct subtraction was
the one spot missing it.

+++ experiments/24-geometric-aug/_models/position_prior_probe.py
@@ -127,7 +127,10 @@ def interaction_ci(labelled: pd.DataFrame, results_a: Path, results_b: Path, *,
         columns={"correctness": "correct_b"})
     merged = labelled.merge(res_a, on="qID", how="inner").merge(
         res_b[["qID", "correct_b"]], on="qID", how="inner")
-    merged["diff"] = merged["correct_b"] - merged["correct_a"]
+    # correctness comes back from results.csv as bool -- numpy/pandas no longer allow
+    # subtracting two bool Series directly ("numpy boolean subtract... not supported"),
+    # so cast explicitly rather than relying on an implicit numeric coercion.
+    merged["diff"] = merged["correct_b"].astype(float) - merged["correct_a"].astype(float)
 
     def _groups(sub: pd.DataFrame) -> dict:
         vk = _video_key(sub)
